refactoring interfaces to proper location - Project & ResearchPaper
