--- src/types/index.ts
@@ -1,4 +1,0 @@
-// Global types index
-// Re-export types from features as needed
-
-// Note: Neural network types are now in @/features/neural-background/types
